requirements.txt
bidict==0.21.2
catkin-pkg==0.4.23
click==8.0.1
distro==1.6.0
docutils==0.17.1
Flask==2.0.1
Flask-PyMongo==2.3.0
Flask-SocketIO==5.1.1
gevent==21.8.0
gevent-websocket==0.10.1
greenlet==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pymongo==3.12.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-engineio==4.2.1
python-socketio==5.4.0
PyYAML==5.4.1
rospkg==1.3.0
Werkzeug==2.0.1
Flask-Cors==3.0.10
